started making abstract calculation class and CP2K class
main idea: calculation base class to provide std api for parsing DFT output files. Then make a CP2K inherited class as proof of concept. NOTE: not sure if this is the best approach to the problem. It might be easier to use a function-based solution (need to discuss more with others).